Convert all seat references to int8_t

+++ Shared/AdditionalPackets.h
@@ -83,7 +83,7 @@ struct tHumanCreatePacket
 {
 	float health;
 	int32_t vehicleNetworkIndex;
-	uint32_t seat;
+	int8_t seat;
 	bool isCrouching;
 	bool isAiming;
 	bool isShooting;
@@ -100,7 +100,7 @@ struct tHumanSyncPacket
 {
 	float health;
 	int32_t vehicleNetworkIndex;
-	uint32_t seat;
+	int8_t seat;
 	bool isCrouching;
 	bool isAiming;
 	bool isShooting;
